ui的处理

+++ PhotonIM/PhotonIM/PhotonIMSource/Components/MenuView/PhotonMenuView.m
@@ -62,7 +62,7 @@
                      [self.menuController setMenuItems:@[copyItem,transpondItem,deleteItem]];
                 }
             }else{
-                if ([item canWithDrawMsg] && [[item userInfo] messageStatus] != PhotonIMMessageStatusFailed) {
+                if ([item canWithDrawMsg] && [[item userInfo] messageStatus] != PhotonIMMessageStatusFailed && [[item userInfo] messageStatus] != PhotonIMMessageStatusSending) {
                      [self.menuController setMenuItems:@[transpondItem,reDrawItem,deleteItem]];
                 }else{
                      [self.menuController setMenuItems:@[transpondItem,deleteItem]];
